refactor: 도메인별 응답 코드 분리 및 마커 인터페이스 적용
