+++ src/main/java/org/sopt/global/api/code/ErrorResultCode.java
@@ -1,0 +1,4 @@
+package org.sopt.global.api.code;
+
+public interface ErrorResultCode extends ResultCode {
+}
