+++ src/main/java/org/sopt/global/api/code/ResultCode.java
@@ -1,0 +1,9 @@
+package org.sopt.global.api.code;
+
+import org.springframework.http.HttpStatus;
+
+public interface ResultCode {
+	HttpStatus getHttpStatus();
+
+	String getMessage();
+}
