+++ src/main/java/org/sopt/global/api/code/SuccessResultCode.java
@@ -1,0 +1,4 @@
+package org.sopt.global.api.code;
+
+public interface SuccessResultCode extends ResultCode {
+}
