+++ src/main/java/org/sopt/global/api/code/member/SuccessCode.java
@@ -1,5 +1,6 @@
-package org.sopt.global.api.code;
+package org.sopt.global.api.code.member;
 
+import org.sopt.global.api.code.SuccessResultCode;
 import org.springframework.http.HttpStatus;
 
 import lombok.Getter;
@@ -7,7 +8,7 @@ import lombok.RequiredArgsConstructor;
 
 @Getter
 @RequiredArgsConstructor
-public enum SuccessCode {
+public enum SuccessCode implements SuccessResultCode {
 	// 200
 	MEMBER_FOUNDED(HttpStatus.OK, "✅ 조회가 완료되었습니다."),
 	MEMBER_DELETED(HttpStatus.OK, "✅ 회원 정보가 삭제되었습니다."),
